🌱 prune dead security subsystems
