+++ changelog.d/changed-5694-prune-dead-security-subsystems.md
@@ -1,0 +1,1 @@
+- Removed the remaining unwired delegation token lifecycle and mint HTTP authenticator/server code while preserving live delegation key publication and agent token minting paths.
